fixed fast calib/placer tests (AutoTracker behaviour), less warns

+++ core/reconstruction/photostationPlacer.cpp
@@ -536,10 +536,10 @@ bool corecvs::PhotostationPlacer::appendAny()
     for (size_t iii = 0; iii < (size_t)speculativity() && iii < scene->placingQueue.size(); ++iii)
     {
         auto fixture = scene->placingQueue[iii];
-        int cnt = scene->getFixtureMatches(scene->placedFixtures, fixture).size();
+        int cnt = (int)scene->getFixtureMatches(scene->placedFixtures, fixture).size();
         auto d3 = scene->getPossibleTracks(fixture);
 		for (auto& t: d3)
-			cnt += std::get<3>(t)->observations__.size();
+            cnt += (int)std::get<3>(t)->observations__.size();
 		if (cnt < shutUpAndAppendMyFixtureInlierThreshold())
 			continue;
 		matches.emplace_back(fixture, cnt);
